Fixed table creation

+++ OdeSolver/Euler.h
@@ -41,7 +41,7 @@ public:
 						rvec,
 						const double&,
 						const double&,
-						const double&,
+						const int&,
 						const OdeFunIF*) override;
 
 	//Get the power of the error
